refactor: fix clippy warnings across codebase
- Fixed 27 clippy warnings (clone_on_copy, unnecessary_cast, field_reassign_with_default, etc.)

- Improved code quality following Rust best practices

- All tests passing: 1411 unit tests, 0 failures

- cargo clippy --all-targets -- -D warnings: PASSED

- Related to Task 11.0 in Epic 17.5

+++ examples/maneuver_nested_flow.rs
@@ -165,7 +165,7 @@ impl PaladinPort for ExamplePaladinPort {
         Ok(PaladinResult {
             output,
             token_count: 150,
-            execution_time_ms: delay as u64,
+            execution_time_ms: delay,
             loop_count: 1,
             stop_reason: StopReason::Completed,
         })
